// src/app/pages/si_alertas/components/si_alertas-table/si_alertas.interface.ts
export interface Si_alertasInterface {
   idsi_alerta?: number;
   si_user_idsi_user?: number;
   tipoAlerta?: string;
   emailRemitente?: string;
   emailDestinatario?: string;
   mensaje?: string;
   vista?: boolean;
   leida?: boolean;
   is_deleted?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}